requirements.txt
Django==5.2.1
DjangoRestFramework==3.14.0
python-dateutil==2.8.2
Gunicorn==21.2.0
psycopg2-binary==2.9.7
celery==5.3.6
redis==5.0.1
python-dotenv==1.1.0
openpyxl==3.1.5
pandas==2.2.3
